backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
sqlalchemy==2.0.29
alembic==1.13.1
httpx==0.27.0
python-multipart==0.0.9
boto3==1.34.84
slowapi==0.1.9
PyJWT==2.8.0
pytest==8.2.2
